Moved Enum name fixes into enum.py.
