+++ generators/google-sheets/sheet2linkml/source/gsheetmodel/enum.py
@@ -96,6 +96,30 @@ class Enum(ModelElement):
         return (self.enum_name or '(unnamed enum)').strip()
 
     @property
+    def fixed_name(self):
+        """
+        :return: The name of this enum, modified to fit the format we use for other enums.
+        """
+        return f'CCDH_{Enum.fix_enum_name(self.name)}'
+
+    @staticmethod
+    def fix_enum_name(enum_name: str):
+        """
+        Transform enum names so that they can be accessed from Python.
+        Ideally, these transformations should be done in the LinkML
+        library somewhere, but for now, we can do that here.
+        """
+
+        # The hyphen in 'CRDC-H' doesn't work properly.
+        fixed_name = re.sub(r'^CRDC-H\.', 'CCDH.', enum_name)
+
+        # The '.'s in the name also mess up the generated Python code.
+        # But we might as well replace everything that isn't alphanumeric.
+        fixed_name = re.sub(r'\W', '_', fixed_name).strip('_')
+
+        return fixed_name
+
+    @property
     def full_name(self) -> str:
         """
         :return: The full name of this enum.
